refactor: move graphql typedef to types
